move all lsp logic to lsp package
